+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2063 - Login security: disable temporarily for 15 mintes user account after 5 wrong attempts
 
 ## [2.23.0] - 2019-03-16
 ### Added